@@ -3,7 +3,7 @@ description: A privacy focused social media based on matrix
 
 # The following line prevents the package from being accidentally published to
 # pub.dev using `pub publish`. This is preferred for private packages.
-publish_to: 'none' # Remove this line if you wish to publish to pub.dev
+publish_to: "none" # Remove this line if you wish to publish to pub.dev
 
 # The following defines the version and build number for your application.
 # A version number is three numbers separated by dots, like 1.2.43
@@ -18,7 +18,7 @@ publish_to: 'none' # Remove this line if you wish to publish to pub.dev
 version: 1.6.0+8
 
 environment:
-  sdk: '>=2.12.0 <3.0.0'
+  sdk: ">=2.12.0 <3.0.0"
 
 dependencies:
   flutter:
@@ -37,17 +37,16 @@ dependencies:
   hive_flutter: ^1.1.0
 
   provider: ^6.0.1
-  
+
   # navigator
-  auto_route: ^4.0.1
+  auto_route: ^5.0.2
 
   # web navigator for SSO
-  webview_flutter: ^3.0.0
+  webview_flutter: ^4.0.1
   url_launcher: ^6.0.3
 
   flutter_typeahead: ^4.0.0
-  
-  
+
   # Post formating
   flutter_markdown:
   emoji_picker_flutter: ^1.2.1
@@ -56,9 +55,7 @@ dependencies:
   image_picker:
   file_picker_cross:
   path_provider: ^2.0.5
-  native_imaging:
-    git: https://gitlab.com/famedly/company/frontend/libraries/native_imaging.git
-
+  native_imaging: ^0.1.0
 
   # time formating
   timeago: ^3.0.2
@@ -67,9 +64,9 @@ dependencies:
   encrypt: ^5.0.1
 
   localstorage: ^4.0.0+1
-  flutter_secure_storage: ^6.0.0
-  
-  # cache 
+  flutter_secure_storage: ^7.0.1
+
+  # cache
   flutter_cache_manager: ^3.0.1
   cached_network_image: ^3.0.0
 
@@ -78,25 +75,29 @@ dependencies:
   cupertino_icons: ^1.0.1
 
   # Notifications
-  flutter_local_notifications: ^9.4.0
+  flutter_local_notifications: ^12.0.2
   unifiedpush: ^4.0.0
   desktop_notifications: ^0.6.3
-  
+
   # Flushbar
   another_flushbar: ^1.10.25
 
   #dialogs
   adaptive_dialog: ^1.2.0
-  
+
   package_info_plus: ^1.0.0
-  
+
   minestrix_chat:
     git:
       url: https://gitlab.com/minestrix/minestrix-chat.git
-    #path: ../minestrix-chat
+    #path: ../minestrix_chat
+  infinite_list:
+    git:
+      url: https://gitlab.com/minestrix/infinite-list.git
+      ref: henri2h/enhancements
+
   collection: ^1.15.0-nullsafety.4
-  f_logs: ^2.0.0
-  connectivity_plus: ^2.3.0
+  connectivity_plus: ^3.0.2
   just_audio: ^0.9.20
   flutter_ringtone_player: ^3.2.0
   sentry: ^6.5.1
@@ -106,13 +107,15 @@ dependencies:
   desktop_webview_window: ^0.1.6
   settings_ui: ^2.0.2
   flutter_datetime_picker: ^1.5.1
+  sentry_flutter: ^6.9.1
+  intl: ^0.18.0
+  scroll_to_index: ^3.0.1
 
-      
 dev_dependencies:
   flutter_test:
     sdk: flutter
-  hive_generator: ^1.1.1
-  auto_route_generator: ^4.0.0
+  hive_generator: ^2.0.0
+  auto_route_generator: ^5.0.2
   build_runner: ^2.1.2
   import_sorter: ^4.6.0
   flutter_lints: ^2.0.1
@@ -121,11 +124,10 @@ flutter:
   uses-material-design: true
 
   assets:
-      - assets/verification.png
-      - assets/sas-emoji.json
-      - assets/icon_512.png
-      - assets/bg_paris.jpg
-  
+    - assets/verification.png
+    - assets/sas-emoji.json
+    - assets/icon_512.png
+    - assets/bg_paris.jpg
 
   fonts:
     - family: Roboto
@@ -138,3 +140,8 @@ flutter:
     - family: NotoEmoji
       fonts:
         - asset: fonts/NotoEmoji/NotoColorEmoji.ttf
+
+dependency_overrides:
+  # Until all dependencies are compatible. Missing: file_picker_cross, flutter_matrix_html
+  ffi: ^2.0.0
+  flutter_svg: ^1.1.5
