@@ -1,7 +1,6 @@
 import 'package:flutter/material.dart';
 import 'package:matrix/matrix.dart';
 import 'package:minestrix_chat/partials/event/matrix_image.dart';
-import 'package:minestrix_chat/utils/social/posts/model/json/social_image_item.dart';
 import 'package:minestrix_chat/utils/social/posts/model/social_item.dart';
 import 'package:minestrix_chat/utils/social/posts/posts_event_extension.dart';
 
@@ -10,8 +9,8 @@ import '../../../partials/post/gallery/post_gallery_nav_button.dart';
 import '../../../partials/post/details/post_header.dart';
 
 class PostGalleryPage extends StatefulWidget {
-  final SocialItem post;
-  final SocialImageItem? image;
+  final Event post;
+  final Event? image;
   final String? selectedImageEventId;
   const PostGalleryPage(
       {Key? key, required this.post, this.image, this.selectedImageEventId})
@@ -30,15 +29,15 @@ class _PostGalleryPageState extends State<PostGalleryPage> {
   Set<Event>? replies;
   Map<Event, dynamic>? nestedReplies;
 
-  String? _replyToMessageId = null;
+  String? _replyToMessageId;
   String? get replyToMessageId =>
       _replyToMessageId ??
-      (modeRef ? selectedImageEventId : widget.post.event?.eventId);
-  Event? get event => modeRef ? _event : widget.post.event;
+      (modeRef ? selectedImageEventId : widget.post.eventId);
+  Event? get event => modeRef ? _event : widget.post;
 
   Future<Timeline> getTimeline() async {
     if (timeline != null) return timeline!;
-    timeline = await widget.post.event?.room.getTimeline(onUpdate: () {
+    timeline = await widget.post.room.getTimeline(onUpdate: () {
       if (timeline != null) {
         loadPost(timeline!);
         if (mounted) setState(() {});
@@ -60,19 +59,14 @@ class _PostGalleryPageState extends State<PostGalleryPage> {
       });
 
   String? selectedImageEventId;
-  SocialImageItem? selectedImage;
+  Event? selectedImage;
   Event? _event;
-  Future<SocialImageItem?>? futureImage;
-  Future<SocialImageItem?> getImage() async {
+  Future<Event?>? futureImage;
+  Future<Event?> getImage() async {
     if (selectedImage != null) return selectedImage;
 
     if (selectedImageEventId != null) {
-      _event =
-          await widget.post.event!.room.getEventById(selectedImageEventId!);
-      if (_event != null) {
-        if (timeline != null) loadPost(timeline!);
-        return SocialImageItem.fromJson(_event!.content);
-      }
+      _event = await widget.post.room.getEventById(selectedImageEventId!);
     }
 
     return null;
@@ -90,7 +84,7 @@ class _PostGalleryPageState extends State<PostGalleryPage> {
 
   void selectImage(int pos) {
     if (!modeRef) {
-      selectedImage = widget.post.images[pos];
+      selectedImage = widget.image;
     } else {
       selectedImageEventId = widget.post.imagesRefEventId[pos];
       _replyToMessageId = null;
@@ -99,17 +93,12 @@ class _PostGalleryPageState extends State<PostGalleryPage> {
     setState(() {});
   }
 
-  int get imgCount => selectedImage != null
-      ? widget.post.images.length
-      : widget.post.imagesRefEventId.length;
-
-  int get pos => selectedImage != null
-      ? widget.post.images.indexOf(selectedImage!)
-      : widget.post.imagesRefEventId.indexOf(selectedImageEventId!);
+  int get imgCount => widget.post.imagesRefEventId.length;
+  int get pos => widget.post.imagesRefEventId.indexOf(selectedImageEventId!);
 
   @override
   Widget build(BuildContext context) {
-    return FutureBuilder<SocialImageItem?>(
+    return FutureBuilder<Event?>(
         future: futureImage,
         builder: (context, snap) {
           return LayoutBuilder(builder: (context, constraints) {
@@ -121,16 +110,18 @@ class _PostGalleryPageState extends State<PostGalleryPage> {
                     fit: StackFit.expand,
                     children: [
                       Builder(builder: (context) {
-                        if (!snap.hasData)
-                          return Center(child: CircularProgressIndicator());
+                        if (!snap.hasData) {
+                          return const Center(
+                              child: CircularProgressIndicator());
+                        }
                         final image = snap.data!;
 
-                        return MatrixSocialImage(
-                            key: Key(image.file?.url ?? pos.toString()),
-                            room: widget.post.event!.room,
-                            image: image,
-                            fit: BoxFit.cover,
-                            borderRadius: BorderRadius.zero);
+                        return MatrixEventImage(
+                          key: Key(image.eventId),
+                          fit: BoxFit.cover,
+                          borderRadius: BorderRadius.zero,
+                          event: widget.post,
+                        );
                       }),
                       PostGalleryNavButton(
                           alignment: Alignment.centerLeft,
@@ -150,7 +141,7 @@ class _PostGalleryPageState extends State<PostGalleryPage> {
                             right: 8,
                             child: MaterialButton(
                                 minWidth: 0,
-                                child: Icon(Icons.close,
+                                child: const Icon(Icons.close,
                                     size: 40, color: Colors.white),
                                 onPressed: () {
                                   Navigator.of(context).pop();
@@ -159,8 +150,8 @@ class _PostGalleryPageState extends State<PostGalleryPage> {
                           bottom: 12,
                           right: 12,
                           child: Text(
-                            (pos + 1).toString() + "/" + imgCount.toString(),
-                            style: TextStyle(
+                            "${pos + 1}/$imgCount",
+                            style: const TextStyle(
                                 fontSize: 26,
                                 color: Colors.white,
                                 fontWeight: FontWeight.w600),
@@ -178,8 +169,7 @@ class _PostGalleryPageState extends State<PostGalleryPage> {
                             child: ListView(
                               children: [
                                 PostHeader(
-                                    event: widget.post.event!,
-                                    allowContext: false),
+                                    event: widget.post, allowContext: false),
                                 if (event != null)
                                   RepliesVue(
                                       timeline: t,
