@@ -1,7 +1,6 @@
 import 'package:flutter/material.dart';
 import 'package:flutter/scheduler.dart';
 import 'package:flutter_datetime_picker/flutter_datetime_picker.dart';
-import 'package:flutter_markdown/flutter_markdown.dart';
 import 'package:matrix/matrix.dart';
 import 'package:minestrix/partials/components/layouts/custom_header.dart';
 import 'package:minestrix/partials/components/minesTrix/MinesTrixTitle.dart';
@@ -10,21 +9,20 @@ import 'package:minestrix/utils/date_time_extension.dart';
 import 'package:minestrix_chat/config/matrix_types.dart';
 import 'package:minestrix_chat/partials/calendar_event/calendar_event_widget.dart';
 import 'package:minestrix_chat/partials/chat/room/room_participants_indicator.dart';
-import 'package:minestrix_chat/partials/chat/settings/conv_settings_card.dart';
-import 'package:minestrix_chat/partials/dialogs/adaptative_dialogs.dart';
+import 'package:minestrix_chat/partials/chat/user/user_selector_dialog.dart';
 import 'package:minestrix_chat/partials/dialogs/custom_dialogs.dart';
 import 'package:minestrix_chat/partials/matrix/matrix_user_item.dart';
 import 'package:minestrix_chat/partials/social/social_gallery_preview_widget.dart';
-import 'package:minestrix_chat/utils/matrix/room_extension.dart';
+import 'package:minestrix_chat/minestrix_chat.dart';
 import 'package:minestrix_chat/utils/poll/poll.dart';
 import 'package:minestrix_chat/utils/social/calendar_events/calendar_events_extension.dart';
 import 'package:minestrix_chat/utils/social/calendar_events/model/calendar_event_model.dart';
 import 'package:minestrix_chat/utils/social/posts/posts_event_extension.dart';
-import 'package:url_launcher/url_launcher.dart';
+import 'package:minestrix_chat/view/room_settings_page.dart';
 
 import '../../partials/components/layouts/layout_view.dart';
+import '../../partials/feed/topic_list_tile.dart';
 import '../../partials/post/post_writer_modal.dart';
-import '../../partials/users/MinesTrixUserSelection.dart';
 
 class CalendarEventPage extends StatefulWidget {
   final Room room;
@@ -82,8 +80,7 @@ class CalendarEventPageState extends State<CalendarEventPage> {
   }
 
   Future<void> inviteUsers() async {
-    List<Profile>? profiles = await AdaptativeDialogs.show(
-        context: context, builder: (a) => MinesTrixUserSelection());
+    List<Profile>? profiles = await MinesTrixUserSelection.show(context);
 
     profiles?.forEach((Profile p) async {
       await widget.room.invite(p.userId);
@@ -99,260 +96,218 @@ class CalendarEventPageState extends State<CalendarEventPage> {
       child: FutureBuilder<Timeline>(
           future: timeline,
           builder: (context, snapT) {
+            final calendarEvent = room.getEventAttendanceEvent();
             return StreamBuilder<Object>(
                 stream: room.onRoomInSync(),
                 builder: (context, snapshot) {
-                  return FutureBuilder<CalendarEvent?>(
-                      future: room.getEventAttendanceEvent(),
-                      builder: (context, snapCalendar) {
-                        final calendarEvent = snapCalendar.data;
-
-                        return LayoutView(
-                            controller: controller,
-                            customHeader: CustomHeader(
-                              title: "Event",
-                              actionButton: [
-                                IconButton(
-                                    icon: const Icon(Icons.settings),
-                                    onPressed: () {
-                                      ConvSettingsCard.show(
-                                          context: context, room: room);
-                                    })
-                              ],
-                            ),
-                            room: room,
-                            headerHeight: 280,
-                            mainWidth: double.infinity,
-                            mainBuilder:
-                                ({required bool displaySideBar}) => Column(
-                                      children: [
-                                        Center(
-                                          child: Padding(
-                                            padding: const EdgeInsets.all(8.0),
-                                            child: Card(
-                                              child: ConstrainedBox(
-                                                constraints:
-                                                    const BoxConstraints(
-                                                        maxWidth: 1200),
-                                                child: Row(
+                  return LayoutView(
+                      controller: controller,
+                      customHeader: CustomHeader(
+                        title: "Event",
+                        actionButton: [
+                          IconButton(
+                              icon: const Icon(Icons.settings),
+                              onPressed: () {
+                                RoomSettingsPage.show(
+                                    context: context, room: room);
+                              })
+                        ],
+                      ),
+                      room: room,
+                      headerHeight: 280,
+                      mainWidth: double.infinity,
+                      mainBuilder: ({required bool displaySideBar}) => Column(
+                            children: [
+                              Center(
+                                child: Padding(
+                                  padding: const EdgeInsets.all(8.0),
+                                  child: Card(
+                                    child: ConstrainedBox(
+                                      constraints:
+                                          const BoxConstraints(maxWidth: 1200),
+                                      child: Row(
+                                        children: [
+                                          Expanded(
+                                            child: Column(
+                                              crossAxisAlignment:
+                                                  CrossAxisAlignment.start,
+                                              children: [
+                                                H2Title(room.name),
+                                                Row(
                                                   children: [
+                                                    const Padding(
+                                                      padding: EdgeInsets.only(
+                                                          left: 10, right: 4),
+                                                      child: Text("By",
+                                                          style: TextStyle(
+                                                              fontSize: 18.0,
+                                                              fontWeight:
+                                                                  FontWeight
+                                                                      .w500)),
+                                                    ),
                                                     Expanded(
-                                                      child: Column(
-                                                        crossAxisAlignment:
-                                                            CrossAxisAlignment
-                                                                .start,
-                                                        children: [
-                                                          H2Title(room.name),
-                                                          Row(
-                                                            children: [
-                                                              const Padding(
-                                                                padding: EdgeInsets
-                                                                    .only(
-                                                                        left:
-                                                                            10,
-                                                                        right:
-                                                                            4),
-                                                                child: Text(
-                                                                    "By",
-                                                                    style: TextStyle(
-                                                                        fontSize:
-                                                                            18.0,
-                                                                        fontWeight:
-                                                                            FontWeight.w500)),
-                                                              ),
-                                                              Expanded(
-                                                                child: MatrixUserItem.fromUser(
-                                                                    room
-                                                                        .creator!,
-                                                                    client: room
-                                                                        .client),
-                                                              ),
-                                                            ],
-                                                          ),
-                                                          Padding(
-                                                            padding:
-                                                                const EdgeInsets
-                                                                    .all(8.0),
-                                                            child: Row(
-                                                              children: [
-                                                                Expanded(
-                                                                  child: RoomParticipantsIndicator(
-                                                                      room:
-                                                                          room),
-                                                                ),
-                                                                ElevatedButton(
-                                                                    onPressed:
-                                                                        inviteUsers,
-                                                                    child: Row(
-                                                                      children: const [
-                                                                        Icon(Icons
-                                                                            .person_add),
-                                                                        SizedBox(
-                                                                            width:
-                                                                                8),
-                                                                        Text(
-                                                                            "Invite"),
-                                                                      ],
-                                                                    ))
-                                                              ],
-                                                            ),
-                                                          )
-                                                        ],
-                                                      ),
+                                                      child: MatrixUserItem
+                                                          .fromUser(
+                                                              room.creator!,
+                                                              client:
+                                                                  room.client),
                                                     ),
                                                   ],
                                                 ),
-                                              ),
+                                                Padding(
+                                                  padding:
+                                                      const EdgeInsets.all(8.0),
+                                                  child: Row(
+                                                    children: [
+                                                      Expanded(
+                                                        child:
+                                                            RoomParticipantsIndicator(
+                                                                room: room),
+                                                      ),
+                                                      ElevatedButton(
+                                                          onPressed:
+                                                              inviteUsers,
+                                                          child: Row(
+                                                            children: const [
+                                                              Icon(Icons
+                                                                  .person_add),
+                                                              SizedBox(
+                                                                  width: 8),
+                                                              Text("Invite"),
+                                                            ],
+                                                          ))
+                                                    ],
+                                                  ),
+                                                )
+                                              ],
                                             ),
                                           ),
-                                        ),
-                                        Padding(
-                                          padding: const EdgeInsets.all(8.0),
-                                          child: Wrap(
-                                            alignment: WrapAlignment.center,
-                                            children: [
-                                              if (snapT.hasData)
-                                                ConstrainedBox(
-                                                    constraints:
-                                                        const BoxConstraints(
-                                                            maxWidth: 600),
-                                                    child: Column(
-                                                      crossAxisAlignment:
-                                                          CrossAxisAlignment
-                                                              .start,
-                                                      children: [
-                                                        Padding(
-                                                          padding:
-                                                              const EdgeInsets
-                                                                  .all(8.0),
-                                                          child: Card(
-                                                            child: Column(
-                                                              crossAxisAlignment:
-                                                                  CrossAxisAlignment
-                                                                      .start,
-                                                              children: [
-                                                                const H2Title(
-                                                                    "About"),
-                                                                (snapT.hasData &&
-                                                                        calendarEvent?.pollId !=
-                                                                            null)
-                                                                    ? FutureBuilder<
-                                                                            Event?>(
-                                                                        future: snapT
-                                                                            .data!
-                                                                            .getEventById(calendarEvent!
-                                                                                .pollId!),
-                                                                        builder:
-                                                                            (context,
-                                                                                snapshot) {
-                                                                          if (snapshot.hasData ==
-                                                                              false) {
-                                                                            return Container();
-                                                                          }
-
-                                                                          Poll p = Poll(
-                                                                              e: snapshot.data!,
-                                                                              t: snapT.data!);
-                                                                          return Row(
-                                                                            mainAxisAlignment:
-                                                                                MainAxisAlignment.center,
-                                                                            children: [
-                                                                              CalendarEventWidget(p: p),
-                                                                            ],
-                                                                          );
-                                                                        })
-                                                                    : ListTile(
-                                                                        leading:
-                                                                            const Icon(Icons
-                                                                                .poll),
-                                                                        title: const Text(
-                                                                            "Set poll id"),
-                                                                        subtitle: Text(room.canSendDefaultStates
-                                                                            ? "Attendance poll could not be found"
-                                                                            : "You don't have the permission to modify this event"),
-                                                                        onTap: !room.canSendDefaultStates
-                                                                            ? null
-                                                                            : () async {
-                                                                                if (calendarEvent != null) {
-                                                                                  await room.setPollAttendance(calendarEvent);
-                                                                                } else {
-                                                                                  await room.setPollAttendance(CalendarEvent());
-                                                                                }
-                                                                              }),
-                                                                Padding(
-                                                                  padding:
-                                                                      const EdgeInsets
-                                                                              .all(
-                                                                          20.0),
-                                                                  child:
-                                                                      MarkdownBody(
-                                                                          data: room
-                                                                              .topic,
-                                                                          onTapLink: (text,
-                                                                              href,
-                                                                              title) async {
-                                                                            if (href !=
-                                                                                null) {
-                                                                              final url = Uri.parse(href);
-                                                                              if (await canLaunchUrl(url)) {
-                                                                                await launchUrl(url);
-                                                                              } else {
-                                                                                throw 'Could not launch $href';
-                                                                              }
-                                                                            }
-                                                                          }),
-                                                                ),
-                                                              ],
-                                                            ),
-                                                          ),
-                                                        ),
-                                                        PostWriterModal(
-                                                            room: room),
-                                                        for (Event e in room
-                                                            .getPosts(
-                                                                snapT.data!,
-                                                                eventTypesFilter: [
-                                                              MatrixTypes.post,
-                                                              EventTypes
-                                                                  .Encrypted,
-                                                              EventTypes
-                                                                  .RoomCreate
-                                                            ]))
-                                                          Post(
-                                                              event: e,
-                                                              onReact:
-                                                                  (Offset e) {},
-                                                              timeline:
-                                                                  snapT.data),
-                                                        if (loading)
-                                                          const PostShimmer(),
-                                                      ],
-                                                    )),
-                                              if (snapT.hasData &&
-                                                  calendarEvent != null)
-                                                ConstrainedBox(
-                                                    constraints:
-                                                        const BoxConstraints(
-                                                            maxWidth: 450),
-                                                    child: Padding(
-                                                      padding:
-                                                          const EdgeInsets.all(
-                                                              8.0),
-                                                      child: CalendarEventInfo(
-                                                          room: room,
-                                                          timeline: snapT.data!,
-                                                          calendarEvent:
-                                                              calendarEvent),
-                                                    )),
-                                            ],
-                                          ),
-                                        ),
-                                      ],
+                                        ],
+                                      ),
                                     ),
-                            headerChildBuilder: (
-                                    {required bool displaySideBar}) =>
-                                Container());
-                      });
+                                  ),
+                                ),
+                              ),
+                              Padding(
+                                padding: const EdgeInsets.all(8.0),
+                                child: Wrap(
+                                  alignment: WrapAlignment.center,
+                                  children: [
+                                    if (snapT.hasData)
+                                      ConstrainedBox(
+                                          constraints: const BoxConstraints(
+                                              maxWidth: 600),
+                                          child: Column(
+                                            crossAxisAlignment:
+                                                CrossAxisAlignment.start,
+                                            children: [
+                                              Padding(
+                                                padding:
+                                                    const EdgeInsets.all(8.0),
+                                                child: Card(
+                                                  child: Column(
+                                                    crossAxisAlignment:
+                                                        CrossAxisAlignment
+                                                            .start,
+                                                    children: [
+                                                      const H2Title("About"),
+                                                      (snapT.hasData &&
+                                                              calendarEvent
+                                                                      ?.pollId !=
+                                                                  null)
+                                                          ? FutureBuilder<
+                                                                  Event?>(
+                                                              future: snapT
+                                                                  .data!
+                                                                  .getEventById(
+                                                                      calendarEvent!
+                                                                          .pollId!),
+                                                              builder: (context,
+                                                                  snapshot) {
+                                                                if (snapshot
+                                                                        .hasData ==
+                                                                    false) {
+                                                                  return Container();
+                                                                }
+
+                                                                Poll p = Poll(
+                                                                    e: snapshot
+                                                                        .data!,
+                                                                    t: snapT
+                                                                        .data!);
+                                                                return Row(
+                                                                  mainAxisAlignment:
+                                                                      MainAxisAlignment
+                                                                          .center,
+                                                                  children: [
+                                                                    CalendarEventWidget(
+                                                                        p: p),
+                                                                  ],
+                                                                );
+                                                              })
+                                                          : ListTile(
+                                                              leading: const Icon(
+                                                                  Icons.poll),
+                                                              title: const Text(
+                                                                  "Set poll id"),
+                                                              subtitle: Text(room
+                                                                      .canSendDefaultStates
+                                                                  ? "Attendance poll could not be found"
+                                                                  : "You don't have the permission to modify this event"),
+                                                              onTap: !room
+                                                                      .canSendDefaultStates
+                                                                  ? null
+                                                                  : () async {
+                                                                      if (calendarEvent !=
+                                                                          null) {
+                                                                        await room
+                                                                            .setPollAttendance(calendarEvent);
+                                                                      } else {
+                                                                        await room
+                                                                            .setPollAttendance(CalendarEvent());
+                                                                      }
+                                                                    }),
+                                                      if (room.canSendDefaultStates ||
+                                                          room.topic.isNotEmpty)
+                                                        TopicListTile(
+                                                            room: room),
+                                                    ],
+                                                  ),
+                                                ),
+                                              ),
+                                              PostWriterModal(room: room),
+                                              for (Event e in room.getPosts(
+                                                  snapT.data!,
+                                                  eventTypesFilter: [
+                                                    MatrixTypes.post,
+                                                    EventTypes.Encrypted,
+                                                    EventTypes.RoomCreate
+                                                  ]))
+                                                Post(
+                                                    event: e,
+                                                    onReact: (Offset e) {},
+                                                    timeline: snapT.data),
+                                              if (loading) const PostShimmer(),
+                                            ],
+                                          )),
+                                    if (snapT.hasData && calendarEvent != null)
+                                      ConstrainedBox(
+                                          constraints: const BoxConstraints(
+                                              maxWidth: 450),
+                                          child: Padding(
+                                            padding: const EdgeInsets.all(8.0),
+                                            child: CalendarEventInfo(
+                                                room: room,
+                                                timeline: snapT.data!,
+                                                calendarEvent: calendarEvent),
+                                          )),
+                                  ],
+                                ),
+                              ),
+                            ],
+                          ),
+                      headerChildBuilder: ({required bool displaySideBar}) =>
+                          Container());
                 });
           }),
     );
