@@ -4,7 +4,7 @@ import 'package:matrix/matrix.dart';
 import 'package:minestrix_chat/utils/matrix_widget.dart';
 import 'package:minestrix/utils/minestrix/minestrix_client_extension.dart';
 import 'package:minestrix_chat/partials/matrix/matrix_image_avatar.dart';
-import 'package:minestrix_chat/utils/matrix/room_extension.dart';
+import 'package:minestrix_chat/minestrix_chat.dart';
 
 class AddUserPage extends StatefulWidget {
   AddUserPage(BuildContext context, {Key? key}) : super(key: key);
