@@ -9,7 +9,7 @@ import 'package:minestrix/partials/components/account/account_card.dart';
 import 'package:minestrix/partials/components/minesTrix/MinesTrixTitle.dart';
 import 'package:minestrix_chat/utils/matrix_widget.dart';
 
-import 'package:minestrix_chat/utils/matrix/room_extension.dart';
+import 'package:minestrix_chat/minestrix_chat.dart';
 
 import '../../../partials/components/layouts/custom_header.dart';
 
