@@ -1,11 +1,15 @@
+import 'dart:async';
+
 import 'package:auto_route/auto_route.dart';
 import 'package:flutter/material.dart';
 import 'package:matrix/matrix.dart';
+// ignore: implementation_imports
+import 'package:matrix/src/utils/cached_stream_controller.dart';
 import 'package:minestrix/pages/minestrix/groups/create_group_page.dart';
-import 'package:minestrix/partials/components/buttons/customTextFutureButton.dart';
+import 'package:minestrix/partials/components/buttons/custom_text_future_button.dart';
 import 'package:minestrix/partials/components/minesTrix/MinesTrixTitle.dart';
-import 'package:minestrix/partials/feed/minestrixProfileNotCreated.dart';
-import 'package:minestrix/partials/feed/notficationBell.dart';
+import 'package:minestrix/partials/feed/minestrix_profile_not_created.dart';
+import 'package:minestrix/partials/feed/notfication_bell.dart';
 import 'package:minestrix/partials/post/post.dart';
 import 'package:minestrix/partials/post/post_writer_modal.dart';
 import 'package:minestrix/router.gr.dart';
@@ -32,49 +36,69 @@ class MinestrixFeed extends StatefulWidget {
 }
 
 class MinestrixFeedState extends State<MinestrixFeed> {
-  Future<List<Event>>? futureEvents;
-  late ScrollController controller;
+  ScrollController controller = ScrollController();
 
+  List<Event>? events;
+  CachedStreamController<int> syncIdStream = CachedStreamController();
+  StreamSubscription? listener;
   String? clientUserId;
 
   void resetPage() {
-    futureEvents = null;
+    isGettingEvents = false;
+    start = 0;
+    _databaseTimelineEvents.clear();
+    events = null;
+    listener?.cancel();
     init();
   }
 
-  void init() {
-    start = 0;
-    _events.clear();
-    isGettingEvents = false;
+  Future<void> loadEvents() async {
+    await Matrix.of(context).client.roomsLoading;
+    await getEvents();
+    syncIdStream.add(syncIdStream.value ?? 0 + 1);
+  }
 
+  void init() {
+    Logs().w("Timeline hooking up");
     final client = Matrix.of(context).client;
 
     clientUserId = client.userID;
 
-    client.onMinestrixUpdate.listen((event) {});
+    listener = client.onNewPost.listen((event) async {
+      Logs().w("Timeline refreshing page");
+      await loadEvents();
+    });
 
-    controller = ScrollController();
-    controller.addListener(onScroll);
+    loadEvents();
   }
 
   @override
   void initState() {
     super.initState();
     init();
+
+    controller.addListener(onScroll);
+  }
+
+  @override
+  void deactivate() {
+    listener?.cancel();
+    super.deactivate();
   }
 
   int start = 0;
-  final List<Event> _events = [];
+  final List<Event> _databaseTimelineEvents = [];
+
   bool isGettingEvents = false;
 
   Future<List<Event>> getEvents() async {
     if (!Settings().optimizedFeed) {
-      return await Matrix.of(context).client.getMinestrixEvents();
+      events = await Matrix.of(context).client.getMinestrixEvents();
+    } else {
+      if (isGettingEvents) return events ?? [];
+      events = await requestEvents();
     }
-    if (_events.isNotEmpty) return _events;
-
-    await requestEvents();
-    return _events;
+    return events!;
   }
 
   Future<List<Event>> requestEvents() async {
@@ -83,8 +107,7 @@ class MinestrixFeedState extends State<MinestrixFeed> {
 
     final client = Matrix.of(context).client;
     try {
-      print("Get events start: $start");
-
+      Logs().w("Request history $start");
       /*
       In order to prevent the application to redraw the feed each time we recieve
       a new post, we make here a copy of the feed and refresh the feed only if the
@@ -94,7 +117,7 @@ class MinestrixFeedState extends State<MinestrixFeed> {
     */
 
       await client.roomsLoading;
-      final events = await client.database?.getEventListFortype(
+      final events = await client.database?.getEventListForType(
               MatrixTypes.post, client.rooms,
               start: start) ??
           [];
@@ -105,11 +128,14 @@ class MinestrixFeedState extends State<MinestrixFeed> {
           .where((element) =>
               element.relationshipType == null && !element.redacted)
           .toList();
-      _events.addAll(fevents);
+
+      // adding events
+      _databaseTimelineEvents.addAll(fevents);
 
       isGettingEvents = false;
-      return fevents;
-    } catch (_) {
+      return _databaseTimelineEvents;
+    } catch (ex, stack) {
+      Logs().e("Timeline fetching error", ex, stack);
       isGettingEvents = false;
       return [];
     }
@@ -134,154 +160,142 @@ class MinestrixFeedState extends State<MinestrixFeed> {
 
           if (client.userID != clientUserId) resetPage();
 
-          return FutureBuilder(
-              future: client.roomsLoading,
-              builder: (context, _) {
-                return FutureBuilder<List<Event>?>(
-                    future: getEvents(),
-                    builder: (context, snap) {
-                      final events = snap.data;
-
-                      if ((events?.length ?? 0) == 0) {
-                        return Column(
-                          mainAxisAlignment: MainAxisAlignment.spaceBetween,
+          return StreamBuilder<int>(
+              stream: syncIdStream.stream,
+              builder: (context, snap) {
+                if (events?.isNotEmpty != true) {
+                  return Column(
+                    mainAxisAlignment: MainAxisAlignment.spaceBetween,
+                    children: [
+                      Expanded(
+                        child: ListView(
                           children: [
-                            Expanded(
-                              child: ListView(
-                                children: [
-                                  const H1Title("Welcome on MinesTRIX"),
-                                  client.prevBatch == null
-                                      ? Column(
-                                          mainAxisAlignment:
-                                              MainAxisAlignment.center,
-                                          children: [
-                                            const MinestrixTitle(),
-                                            SyncStatusCard(client: client),
-                                          ],
-                                        )
-                                      : Column(
-                                          children: [
-                                            H2Title(
-                                                client.userRoomCreated != true
-                                                    ? "First time here ?"
-                                                    : "Your timeline is empty"),
-                                            if (client.userRoomCreated != true)
-                                              const Padding(
-                                                padding: EdgeInsets.all(8.0),
-                                                child:
-                                                    MinestrixProfileNotCreated(),
-                                              ),
-                                            if (client.userRoomCreated == true)
-                                              Padding(
-                                                padding:
-                                                    const EdgeInsets.all(8.0),
-                                                child: CustomTextFutureButton(
-                                                    icon: const Icon(
-                                                        Icons.post_add),
-                                                    text:
-                                                        "Write your first post",
-                                                    onPressed: () async {
-                                                      context.pushRoute(
-                                                          PostEditorRoute());
-                                                    }),
-                                              ),
-                                            Padding(
-                                              padding:
-                                                  const EdgeInsets.all(8.0),
-                                              child: CustomTextFutureButton(
-                                                  icon: const Icon(
-                                                      Icons.group_add),
-                                                  text: "Create a group",
-                                                  onPressed: () async {
-                                                    AdaptativeDialogs.show(
-                                                        context: context,
-                                                        builder: (context) =>
-                                                            const CreateGroupPage());
-                                                  }),
-                                            ),
-                                          ],
+                            const H1Title("Welcome on MinesTRIX"),
+                            client.prevBatch == null
+                                ? Column(
+                                    mainAxisAlignment: MainAxisAlignment.center,
+                                    children: [
+                                      const MinestrixTitle(),
+                                      SyncStatusCard(client: client),
+                                    ],
+                                  )
+                                : Column(
+                                    children: [
+                                      H2Title(events == null
+                                          ? "First time here ?"
+                                          : "Your timeline is empty"),
+                                      if (events != null)
+                                        const Padding(
+                                          padding: EdgeInsets.all(8.0),
+                                          child: MinestrixProfileNotCreated(),
                                         ),
-                                ],
-                              ),
-                            ),
-                            Padding(
-                              padding: const EdgeInsets.only(
-                                bottom: 50,
-                                left: 8.0,
-                                right: 8,
-                              ),
-                              child: CustomTextFutureButton(
-                                  icon: const Icon(Icons.refresh),
-                                  text: "Refresh rooms",
-                                  onPressed: () async {
-                                    setState(() {});
-                                  }),
-                            ),
-                          ],
-                        );
-                      }
-
-                      return Column(
-                        children: [
-                          CustomHeader(
-                              title: "Feed",
-                              actionButton: [
-                                IconButton(
-                                    icon: const Icon(Icons.group_add),
-                                    onPressed: () {
-                                      AdaptativeDialogs.show(
-                                          context: context,
-                                          builder: (context) =>
-                                              const CreateGroupPage());
-                                    }),
-                                const NotificationBell()
-                              ],
-                              child: MaterialButton(
-                                  minWidth: 0,
-                                  shape: const CircleBorder(),
-                                  child:
-                                      AvatarBottomBar(key: Key(client.userID!)),
-                                  onPressed: () {
-                                    context.pushRoute(UserViewRoute(
-                                        userID:
-                                            Matrix.of(context).client.userID));
-                                  })),
-                          Expanded(
-                            child: CustomListViewWithEmoji(
-                                itemCount: events!.length + 1,
-                                controller: controller,
-                                itemBuilder: (BuildContext c, int i,
-                                    void Function(Offset, Event) onReact) {
-                                  if (i == 0) {
-                                    return Column(
-                                      crossAxisAlignment:
-                                          CrossAxisAlignment.start,
-                                      children: [
+                                      if (client.userRoomCreated == true)
                                         Padding(
-                                          padding: const EdgeInsets.symmetric(
-                                              vertical: 8.0),
-                                          child: StoriesList(client: client),
+                                          padding: const EdgeInsets.all(8.0),
+                                          child: CustomTextFutureButton(
+                                              icon: const Icon(Icons.post_add),
+                                              text: "Write your first post",
+                                              onPressed: () async {
+                                                context.pushRoute(
+                                                    PostEditorRoute(
+                                                        room: client
+                                                            .minestrixUserRoom));
+                                              }),
                                         ),
-                                        if (client.minestrixUserRoom.isNotEmpty)
-                                          PostWriterModal(), // TODO: set the actual rom we are displaying
-                                        const FriendSuggestionsList(),
-                                      ],
-                                    );
-                                  }
-
-                                  return Padding(
-                                      padding: const EdgeInsets.symmetric(
-                                        vertical: 2,
+                                      Padding(
+                                        padding: const EdgeInsets.all(8.0),
+                                        child: CustomTextFutureButton(
+                                            icon: const Icon(Icons.group_add),
+                                            text: "Create a group",
+                                            onPressed: () async {
+                                              AdaptativeDialogs.show(
+                                                  context: context,
+                                                  builder: (context) =>
+                                                      const CreateGroupPage());
+                                            }),
                                       ),
-                                      child: Post(
-                                          event: events[i - 1],
-                                          onReact: (Offset e) =>
-                                              onReact(e, events[i - 1])));
-                                }),
-                          ),
+                                    ],
+                                  ),
+                          ],
+                        ),
+                      ),
+                      Padding(
+                        padding: const EdgeInsets.only(
+                          bottom: 50,
+                          left: 8.0,
+                          right: 8,
+                        ),
+                        child: CustomTextFutureButton(
+                            icon: const Icon(Icons.refresh),
+                            text: "Refresh rooms",
+                            onPressed: () async {
+                              getEvents();
+                            }),
+                      ),
+                    ],
+                  );
+                }
+
+                Logs().w("Timeline build page");
+                return Column(
+                  children: [
+                    CustomHeader(
+                        title: "Feed",
+                        actionButton: [
+                          IconButton(
+                              icon: const Icon(Icons.group_add),
+                              onPressed: () {
+                                AdaptativeDialogs.show(
+                                    context: context,
+                                    builder: (context) =>
+                                        const CreateGroupPage());
+                              }),
+                          const NotificationBell()
                         ],
-                      );
-                    });
+                        child: MaterialButton(
+                            minWidth: 0,
+                            shape: const CircleBorder(),
+                            child: AvatarBottomBar(key: Key(client.userID!)),
+                            onPressed: () {
+                              context.pushRoute(UserViewRoute(
+                                  userID: Matrix.of(context).client.userID));
+                            })),
+                    Expanded(
+                      child: CustomListViewWithEmoji(
+                          itemCount: events!.length + 1,
+                          controller: controller,
+                          itemBuilder: (BuildContext c, int i,
+                              void Function(Offset, Event) onReact) {
+                            if (i == 0) {
+                              return Column(
+                                crossAxisAlignment: CrossAxisAlignment.start,
+                                children: [
+                                  Padding(
+                                    padding: const EdgeInsets.symmetric(
+                                        vertical: 8.0),
+                                    child: StoriesList(client: client),
+                                  ),
+                                  if (client.minestrixUserRoom.isNotEmpty)
+                                    const PostWriterModal(), // TODO: set the actual rom we are displaying
+                                  const FriendSuggestionsList(),
+                                ],
+                              );
+                            }
+
+                            return Padding(
+                                padding: const EdgeInsets.symmetric(
+                                  vertical: 2,
+                                ),
+                                child: Post(
+                                    event: events![i - 1],
+                                    key: Key(events![i - 1].eventId +
+                                        events![i - 1].status.toString()),
+                                    onReact: (Offset e) =>
+                                        onReact(e, events![i - 1])));
+                          }),
+                    ),
+                  ],
+                );
               });
         });
   }
