@@ -1,7 +1,7 @@
 import 'package:flutter/material.dart';
 import 'package:matrix/matrix.dart';
 
-import 'package:minestrix/partials/components/buttons/customFutureButton.dart';
+import 'package:minestrix/partials/components/buttons/custom_future_button.dart';
 import 'package:minestrix_chat/utils/matrix_widget.dart';
 import 'package:minestrix/utils/minestrix/minestrix_client_extension.dart';
 
@@ -27,6 +27,7 @@ class MinestrixProfileNotCreated extends StatelessWidget {
               icon: Icon(Icons.skateboarding_outlined,
                   color: Theme.of(context).colorScheme.onPrimary),
               color: Theme.of(context).primaryColor,
+              onPressed: sclient.createPrivateMinestrixProfile,
               children: [
                 Text("Create your private MinesTRIX profile",
                     style: TextStyle(
@@ -37,8 +38,7 @@ class MinestrixProfileNotCreated extends StatelessWidget {
                     style: TextStyle(
                         fontSize: 14,
                         color: Theme.of(context).colorScheme.onPrimary)),
-              ],
-              onPressed: sclient.createPrivateMinestrixProfile);
+              ]);
         });
   }
 }
