@@ -6,21 +6,21 @@ class MinestrixTitle extends StatelessWidget {
   @override
   Widget build(BuildContext context) {
     return Padding(
-      padding: EdgeInsets.all(50),
+      padding: const EdgeInsets.all(50),
       child: Column(mainAxisSize: MainAxisSize.min, children: [
         Row(
           children: [
             Image.asset("assets/icon_512.png",
                 width: 72, height: 72, cacheWidth: 140),
-            SizedBox(width: 28),
+            const SizedBox(width: 28),
             Expanded(
               child: Column(
                 crossAxisAlignment: CrossAxisAlignment.start,
                 children: [
-                  Text("MinesTRIX",
+                  const Text("MinesTRIX",
                       style:
                           TextStyle(fontSize: 28, fontWeight: FontWeight.w800)),
-                  Text("A privacy focused social media based on MATRIX",
+                  const Text("A privacy focused social media based on MATRIX",
                       style: TextStyle(fontSize: 16))
                 ],
               ),
