@@ -7,7 +7,6 @@ import 'package:minestrix/router.gr.dart';
 import 'package:minestrix_chat/config/matrix_types.dart';
 import 'package:minestrix_chat/partials/event/matrix_image.dart';
 import 'package:minestrix_chat/partials/feed/posts/matrix_post_content.dart';
-import 'package:minestrix_chat/utils/social/posts/model/json/social_image_item.dart';
 
 class PostContent extends StatelessWidget {
   final Event event;
@@ -26,14 +25,15 @@ class PostContent extends StatelessWidget {
     switch (event.type) {
       case EventTypes.Encrypted:
       case MatrixTypes.post:
+      case MatrixTypes.comment:
         return MatrixPostContent(
             event: event,
             imageMaxHeight: imageMaxHeight,
             imageMaxWidth: imageMaxWidth,
             disablePadding: disablePadding,
-            onImagePressed: (post, {SocialImageItem? image, String? ref}) {
+            onImagePressed: (Event post, {Event? imageEvent, String? ref}) {
               context.pushRoute(PostGalleryRoute(
-                  image: image, post: post, selectedImageEventId: ref));
+                  image: imageEvent, post: post, selectedImageEventId: ref));
             });
 
       case EventTypes.Message:
