@@ -2,8 +2,7 @@ import 'package:collection/collection.dart';
 import 'package:matrix/matrix.dart';
 import 'package:minestrix_chat/config/matrix_types.dart';
 import 'package:minestrix_chat/utils/matrix/power_levels_extension.dart';
-import 'package:minestrix_chat/utils/room_feed_extension.dart';
-import 'package:minestrix_chat/utils/matrix/room_extension.dart';
+import 'package:minestrix_chat/minestrix_chat.dart';
 import 'package:minestrix_chat/utils/social/posts/posts_event_extension.dart';
 
 extension MinestrixClientExtension on Client {
@@ -130,8 +129,12 @@ extension MinestrixClientExtension on Client {
 
   Iterable<Event> getPostReactions(Iterable<Event> events) => events
       .where((e) =>
-          [MatrixTypes.post, EventTypes.Message, EventTypes.Encrypted]
-              .contains(e.type) &&
+          [
+            MatrixTypes.post,
+            MatrixTypes.comment,
+            EventTypes.Message,
+            EventTypes.Encrypted
+          ].contains(e.type) &&
           !e.redacted)
       .toList();
 
