@@ -1,10 +1,27 @@
 import 'dart:async';
 
 import 'package:matrix/matrix.dart';
+import 'package:minestrix_chat/config/matrix_types.dart';
 import 'minestrix_client_extension.dart';
 
 extension MinestrixNotifications on Client {
-  Stream get onMinestrixUpdate => onSync.stream.where((up) => up.hasRoomUpdate);
+  Stream<SyncUpdate> get onMinestrixUpdate =>
+      onSync.stream.where((up) => up.hasRoomUpdate);
+
+  Stream<SyncUpdate> get onNewPost => onSync.stream.where((up) {
+        if (up.rooms?.join?.values != null) {
+          for (var room in up.rooms!.join!.values) {
+            if (room.timeline?.events?.isNotEmpty == true) {
+              for (var event in room.timeline!.events!) {
+                if (event.type == MatrixTypes.post) {
+                  return true;
+                }
+              }
+            }
+          }
+        }
+        return false;
+      });
 
   /// This method should fetch all the user events and build the notification feed
   // TODO : add a way to check if the user has been mentionned or not.
@@ -16,7 +33,7 @@ extension MinestrixNotifications on Client {
     minestrixInvites.forEach((room) {
       Notification n = Notification();
       n.body = "Friend request";
-      n.type = NotificationType.FriendRequest;
+      n.type = NotificationType.friendRequest;
       notifications.add(n);
     });
     return notifications;
@@ -24,7 +41,7 @@ extension MinestrixNotifications on Client {
 
   get hasNotifications => _hasNotifications();
   bool _hasNotifications() {
-    if (notifications.length == 0) return false;
+    if (notifications.isEmpty) return false;
     return true;
   }
 
@@ -47,4 +64,4 @@ class Notification {
   Profile? p;
 }
 
-enum NotificationType { Text, FriendRequest }
+enum NotificationType { Text, friendRequest }
