@@ -1,6 +1,5 @@
 import 'package:matrix/matrix.dart';
-import 'package:minestrix_chat/utils/matrix/room_extension.dart';
-import 'package:minestrix_chat/utils/room_feed_extension.dart';
+import 'package:minestrix_chat/minestrix_chat.dart';
 
 class MinestrixRoom {
   final Room room;
